# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AstraSynq API",
    "description": "Server sessions and workspace RBAC. Login requires /auth/csrf and X-CSRF-Token; mutations require the token returned by login/me.",
    "version": "0.5.0-rc.1"
  },
  "paths": {
...
    "/api/v1/dashboard/summary": {
      "get": {
        "tags": [
          "dashboard"
        ],
        "summary": "Summary",
        "operationId": "summary_api_v1_dashboard_summary_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Summary"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          },
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
...
    "/api/v1/imports": {
      "post": {
        "tags": [
          "imports"
        ],
        "summary": "Upload",
        "operationId": "upload_api_v1_imports_post",
        "security": [
          {
            "APIKeyCookie": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_api_v1_imports_post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportResult"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          }
        }
      },
      "get": {
        "tags": [
          "imports"
        ],
        "summary": "List Imports",
        "operationId": "list_imports_api_v1_imports_get",
        "security": [
          {
            "APIKeyCookie": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "uploaded",
                    "mapped",
                    "analyzed",
                    "completed"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 120,
              "default": "",
              "title": "Search"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/imports/{import_id}": {
      "get": {
        "tags": [
          "imports"
        ],
        "summary": "Get Import",
        "operationId": "get_import_api_v1_imports__import_id__get",
        "security": [
          {
            "APIKeyCookie": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "import_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Import Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportResult"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          }
        }
      }
    },
...
    "/api/v1/imports/{import_id}/analyze": {
      "post": {
        "tags": [
          "imports"
        ],
        "summary": "Analyze",
        "operationId": "analyze_api_v1_imports__import_id__analyze_post",
        "security": [
          {
            "APIKeyCookie": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "import_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Import Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportResult"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          }
        }
      }
    },
    "/api/v1/imports/{import_id}/commit": {
      "post": {
        "tags": [
          "imports"
        ],
        "summary": "Commit",
        "operationId": "commit_api_v1_imports__import_id__commit_post",
        "security": [
          {
            "APIKeyCookie": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "import_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Import Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommitResult"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          }
        }
      }
    },
    "/api/v1/imports/{import_id}/issues.csv": {
      "get": {
        "tags": [
          "imports"
        ],
        "summary": "Issues",
        "operationId": "issues_api_v1_imports__import_id__issues_csv_get",
        "security": [
          {
            "APIKeyCookie": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "import_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Import Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/leads/export.csv": {
      "get": {
        "tags": [
          "leads"
        ],
        "summary": "Export",
        "operationId": "export_api_v1_leads_export_csv_get",
        "security": [
          {
            "APIKeyCookie": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 120,
              "default": "",
              "title": "Search"
            }
          },
          {
            "name": "company",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 120,
              "default": "",
              "title": "Company"
            }
          },
          {
            "name": "import_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Import Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/leads": {
      "get": {
        "tags": [
          "leads"
        ],
        "summary": "List Leads",
        "operationId": "list_leads_api_v1_leads_get",
        "security": [
          {
            "APIKeyCookie": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 120,
              "default": "",
              "title": "Search"
            }
          },
          {
            "name": "company",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 120,
              "default": "",
              "title": "Company"
            }
          },
          {
            "name": "import_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Import Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeadPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_api_v1_imports_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_api_v1_imports_post"
      },
...
      "ErrorResponse": {
        "properties": {
          "detail": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ImportPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Activity"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "ImportPage"
      },
      "ImportResult": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "status": {
            "type": "string",
            "enum": [
              "uploaded",
              "mapped",
              "analyzed",
              "completed"
            ],
            "title": "Status"
          },
          "columns": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Columns"
          },
          "mapping": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Mapping"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "valid": {
            "type": "integer",
            "title": "Valid"
          },
          "invalid": {
            "type": "integer",
            "title": "Invalid"
          },
          "duplicate": {
            "type": "integer",
            "title": "Duplicate"
          },
          "inserted": {
            "type": "integer",
            "title": "Inserted",
            "default": 0
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/PreviewRow"
            },
            "type": "array",
            "title": "Rows"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "filename",
          "status",
          "columns",
          "mapping",
          "total",
          "valid",
          "invalid",
          "duplicate",
          "rows",
          "created_at"
        ],
        "title": "ImportResult"
      },
...
      "Summary": {
        "properties": {
          "integration_count": {
            "type": "integer",
            "title": "Integration Count",
            "default": 0
          },
          "record_count": {
            "type": "integer",
            "title": "Record Count"
          },
          "import_count": {
            "type": "integer",
            "title": "Import Count"
          },
          "valid": {
            "type": "integer",
            "title": "Valid"
          },
          "invalid": {
            "type": "integer",
            "title": "Invalid"
          },
          "duplicate": {
            "type": "integer",
            "title": "Duplicate"
          },
          "quality": {
            "type": "number",
            "title": "Quality"
          },
          "series": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Series"
          },
          "imports": {
            "items": {
              "$ref": "#/components/schemas/Activity"
            },
            "type": "array",
            "title": "Imports"
          },
          "records": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Records"
          },
          "mode": {
            "type": "string",
            "title": "Mode",
            "default": "postgresql"
          }
        },
        "type": "object",
        "required": [
          "record_count",
          "import_count",
          "valid",
          "invalid",
          "duplicate",
          "quality",
          "series",
          "imports",
          "records"
        ],
        "title": "Summary"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token"
      },
      "APIKeyCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "astrasynq_session"
      }
    }
  }
}